Cleaning

+++ ngrok_deploy.py
@@ -12,15 +12,24 @@ time.sleep(3)
 r = requests.get("http://localhost:4040/api/tunnels")
 ngrok_link = r.json()["tunnels"][0]["public_url"]
 
+
 payload = {"apiKey": TIMEULAR_KEY, "apiSecret": TIMEULAR_SECRET}
 r = requests.post("https://api.timeular.com/api/v3/developer/sign-in", json=payload)
+
+if r.status_code != 200:
+    print(r.text)
+    sys.exit(-1)
+
 BEARER = r.json()["token"]
 
 r = requests.delete(
     "https://api.timeular.com/api/v3/webhooks/subscription",
     headers={"Authorization": "Bearer " + BEARER},
 )
-print(r.text)
+
+if r.status_code != 200:
+    print(r.text)
+    sys.exit(-1)
 
 for event in ["trackingStarted", "trackingStopped"]:
     payload = {"target_url": ngrok_link + "/" + event, "event": event}
@@ -29,4 +38,6 @@ for event in ["trackingStarted", "trackingStopped"]:
         json=payload,
         headers={"Authorization": "Bearer " + BEARER},
     )
-    print(r.text)
+    if r.status_code != 200:
+        print(r.text)
+        sys.exit(-1)
